# frontend/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Baltiyskiy Bereg Backend API",
    "description": "Admin API for the Baltiyskiy Bereg service-desk assistant. Use it for admin authentication, dashboard analytics, appeals management, and assistant settings.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/admin/auth/login": {
      "post": {
        "tags": [
          "Admin Auth"
        ],
        "summary": "Login to admin panel",
        "description": "Authenticate an admin user and return a bearer token for protected endpoints.",
        "operationId": "login_api_v1_admin_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/me": {
      "get": {
        "tags": [
          "Admin Auth"
        ],
        "summary": "Get current admin profile",
        "description": "Return the authenticated admin identity resolved from the bearer token.",
        "operationId": "me_api_v1_admin_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/dashboard/summary": {
      "get": {
        "tags": [
          "Admin Dashboard"
        ],
        "summary": "Get dashboard summary",
        "description": "Return top-level health and support efficiency metrics for the selected time period.",
        "operationId": "summary_api_v1_admin_dashboard_summary_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "24h",
              "title": "Period"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/dashboard/messages-timeseries": {
      "get": {
        "tags": [
          "Admin Dashboard"
        ],
        "summary": "Get messages time series",
        "description": "Return aggregated message activity buckets for the selected time period.",
        "operationId": "messages_timeseries_api_v1_admin_dashboard_messages_timeseries_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "24h",
              "title": "Period"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagesTimeseriesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/appeals": {
      "get": {
        "tags": [
          "Admin Appeals"
        ],
        "summary": "List appeals",
        "description": "Return appeals filtered by date range, scope, and status for the admin queue.",
        "operationId": "list_appeals_api_v1_admin_appeals_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "all",
              "title": "Scope"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppealsListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/appeals/{appeal_id}": {
      "get": {
        "tags": [
          "Admin Appeals"
        ],
        "summary": "Get appeal details",
        "description": "Return a single appeal with its current status and timestamps.",
        "operationId": "get_appeal_api_v1_admin_appeals__appeal_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "appeal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Appeal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppealDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/appeals/{appeal_id}/take": {
      "post": {
        "tags": [
          "Admin Appeals"
        ],
        "summary": "Take appeal into work",
        "description": "Assign the appeal to the current admin and move it into in-progress state.",
        "operationId": "take_appeal_api_v1_admin_appeals__appeal_id__take_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "appeal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Appeal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppealStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/appeals/{appeal_id}/close": {
      "post": {
        "tags": [
          "Admin Appeals"
        ],
        "summary": "Close appeal",
        "description": "Mark the appeal as closed and return its closure timestamp.",
        "operationId": "close_appeal_api_v1_admin_appeals__appeal_id__close_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "appeal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Appeal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppealCloseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/appeals/{appeal_id}/rating-request": {
      "post": {
        "tags": [
          "Admin Appeals"
        ],
        "summary": "Send rating request",
        "description": "Mark that the appeal customer satisfaction request has been sent.",
        "operationId": "send_rating_request_api_v1_admin_appeals__appeal_id__rating_request_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "appeal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Appeal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppealRatingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/settings": {
      "get": {
        "tags": [
          "Admin Settings"
        ],
        "summary": "Get assistant settings",
        "description": "Return the current assistant runtime settings used by the admin panel and backend orchestration.",
        "operationId": "get_settings_api_v1_admin_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantSettingsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Admin Settings"
        ],
        "summary": "Update assistant settings",
        "description": "Persist assistant runtime settings such as tone of voice, confidence threshold, and top-k retrieval.",
        "operationId": "update_settings_api_v1_admin_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantSettingsUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantSettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AdminMeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "login": {
            "type": "string",
            "title": "Login"
          }
        },
        "type": "object",
        "required": [
          "id",
          "login"
        ],
        "title": "AdminMeResponse"
      },
      "AppealCloseResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "const": "closed",
            "title": "Status"
          },
          "closed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Closed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "closed_at"
        ],
        "title": "AppealCloseResponse"
      },
      "AppealDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "employee_login": {
            "type": "string",
            "title": "Employee Login"
          },
          "scope": {
            "type": "string",
            "enum": [
              "ticket",
              "assistant"
            ],
            "title": "Scope"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "in_progress",
              "closed"
            ],
            "title": "Status"
          },
          "priority": {
            "type": "string",
            "enum": [
              "p1",
              "p2",
              "p3",
              "p4"
            ],
            "title": "Priority"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "processing_duration_minutes": {
            "type": "integer",
            "title": "Processing Duration Minutes"
          },
          "closed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "csat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Csat"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "employee_login",
          "scope",
          "status",
          "priority",
          "category",
          "processing_duration_minutes",
          "closed_at",
          "csat",
          "created_at"
        ],
        "title": "AppealDetailResponse"
      },
      "AppealListItemResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "employee_login": {
            "type": "string",
            "title": "Employee Login"
          },
          "scope": {
            "type": "string",
            "enum": [
              "ticket",
              "assistant"
            ],
            "title": "Scope"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "in_progress",
              "closed"
            ],
            "title": "Status"
          },
          "priority": {
            "type": "string",
            "enum": [
              "p1",
              "p2",
              "p3",
              "p4"
            ],
            "title": "Priority"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "processing_duration_minutes": {
            "type": "integer",
            "title": "Processing Duration Minutes"
          },
          "closed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "csat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Csat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "employee_login",
          "scope",
          "status",
          "priority",
          "category",
          "processing_duration_minutes",
          "closed_at",
          "csat"
        ],
        "title": "AppealListItemResponse"
      },
      "AppealRatingResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "rating_request_sent": {
            "type": "boolean",
            "title": "Rating Request Sent"
          }
        },
        "type": "object",
        "required": [
          "id",
          "rating_request_sent"
        ],
        "title": "AppealRatingResponse"
      },
      "AppealStatusResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "const": "in_progress",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status"
        ],
        "title": "AppealStatusResponse"
      },
      "AppealsListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AppealListItemResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "AppealsListResponse"
      },
      "AssistantSettingsResponse": {
        "properties": {
          "tone_of_voice": {
            "type": "string",
            "title": "Tone Of Voice"
          },
          "confidence_threshold": {
            "type": "number",
            "title": "Confidence Threshold"
          },
          "top_k": {
            "type": "integer",
            "title": "Top K"
          },
          "use_articles": {
            "type": "boolean",
            "title": "Use Articles"
          }
        },
        "type": "object",
        "required": [
          "tone_of_voice",
          "confidence_threshold",
          "top_k",
          "use_articles"
        ],
        "title": "AssistantSettingsResponse"
      },
      "AssistantSettingsUpdateRequest": {
        "properties": {
          "tone_of_voice": {
            "type": "string",
            "minLength": 1,
            "title": "Tone Of Voice"
          },
          "confidence_threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence Threshold"
          },
          "top_k": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Top K"
          },
          "use_articles": {
            "type": "boolean",
            "title": "Use Articles"
          }
        },
        "type": "object",
        "required": [
          "tone_of_voice",
          "confidence_threshold",
          "top_k",
          "use_articles"
        ],
        "title": "AssistantSettingsUpdateRequest"
      },
      "DashboardSummaryResponse": {
        "properties": {
          "health": {
            "type": "string",
            "enum": [
              "healthy",
              "degraded",
              "down"
            ],
            "title": "Health"
          },
          "open_appeals_count": {
            "type": "integer",
            "title": "Open Appeals Count"
          },
          "in_progress_appeals_count": {
            "type": "integer",
            "title": "In Progress Appeals Count"
          },
          "assistant_resolution_rate": {
            "type": "number",
            "title": "Assistant Resolution Rate"
          },
          "avg_resolution_minutes": {
            "type": "number",
            "title": "Avg Resolution Minutes"
          },
          "csat_avg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Csat Avg"
          }
        },
        "type": "object",
        "required": [
          "health",
          "open_appeals_count",
          "in_progress_appeals_count",
          "assistant_resolution_rate",
          "avg_resolution_minutes",
          "csat_avg"
        ],
        "title": "DashboardSummaryResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "MessagesTimeseriesPoint": {
        "properties": {
          "ts": {
            "type": "string",
            "format": "date-time",
            "title": "Ts"
          },
          "messages_count": {
            "type": "integer",
            "title": "Messages Count"
          }
        },
        "type": "object",
        "required": [
          "ts",
          "messages_count"
        ],
        "title": "MessagesTimeseriesPoint"
      },
      "MessagesTimeseriesResponse": {
        "properties": {
          "period": {
            "type": "string",
            "enum": [
              "1h",
              "6h",
              "24h",
              "7d",
              "30d",
              "90d"
            ],
            "title": "Period"
          },
          "total_messages": {
            "type": "integer",
            "title": "Total Messages"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/MessagesTimeseriesPoint"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "period",
          "total_messages",
          "points"
        ],
        "title": "MessagesTimeseriesResponse"
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "TokenResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}